Replaces ints by bytes for polydraw attributes

+++ Server/internal/services/potrace/model/xmlsvg.go
@@ -30,10 +30,10 @@ type XMLPath struct {
 	D         string    `xml:"d,attr"`
 	Time      int       `xml:"http://example.org/polydraw time,attr"`
 	Order     int       `xml:"http://example.org/polydraw order,attr"`
-	Color     int       `xml:"http://example.org/polydraw color,attr"`
+	Color     byte      `xml:"http://example.org/polydraw color,attr"`
 	Eraser    bool      `xml:"http://example.org/polydraw eraser,attr"`
 	Brush     string    `xml:"http://example.org/polydraw brush,attr"`
-	BrushSize int       `xml:"http://example.org/polydraw brushsize,attr"`
+	BrushSize byte      `xml:"http://example.org/polydraw brushsize,attr"`
 	ID        uuid.UUID `xml:"id,attr"`
 
 	FirstCommand svgparser.Command `xml:"-"`
